fix: rename import Command to CommandName
