+++ src/cli/cli-application.ts
@@ -1,6 +1,6 @@
 import { ICommand } from './commands/command.interface.js';
 import { CommandParser } from './command-parser.js';
-import { CommandName } from '../shared/types/command.enum.js';
+import { CommandName } from '../shared/types/command-name.enum.js';
 
 type CommandCollection = Record<string, ICommand>;
