+++ src/shared/types/index.ts
@@ -1,7 +1,7 @@
 export * from './geo.type.js';
 export * from './user.type.js';
 export * from './user-type.enum.js';
-export * from './command.enum.js';
+export * from './command-name.enum.js';
 export * from './comment.type.js';
 export * from './housing-type.enum.js';
 export * from './housing-feature.enum.js';
